backend/requirements.txt
fastapi
uvicorn
httpx
jinja2
python-dotenv
openai
anthropic
pinecone
pyyaml 